Add Github Actions to run tests and build docs
- Remove travis CI config
- Remove requirements*_raw.txt
- Remove runcodecheckers script
- flake8 cleanup of codebase
- Add `.python-version` file, run code on python 3.7.13
- Add logging to various files
- Bump rubberband version to 1.0.0
- Add additional installation requirements to README
- Add a test for compare view
- Bugfix `debug` user permissions issue
- Bugfix firafonts css loading

Though the github actions are failing, following these steps I was able to get the tests running and passing locally. My best guess as to what is happening here is that the elasticsearch service is using the incorrect version of java, and hence having a problem booting up. I'll merge this with the  broken actions, and then follow up with the first elasticsearch upgrade (5.0.0), which should hopefully resolve the github actions issues since I'll then be able to use this docker container in the action instead of trying to run it as a service: https://www.docker.elastic.co/r/elasticsearch/elasticsearch:5.0.0