Upgrade to Elasticsearch 7.0.0
Elasticsearch has deprecated the notion of `doc_types`, and now recommends putting different document types in different indices[1]. As such, I've created indices for the new `File`, `Settings`, and `Result` models. The values for different settings have inconsistent types, so settings needs to be in their own index so that we can turn on the `index.mapping.ignore_malformed` setting on that index [2]. This is done in the `bin/rubberband-ctl` in the `create-indices` function. Also in this function, we are setting the `index.mapping.total_fields.limit` value to 5000 on the `Settings` and `Results` indices. The default, which is 1000, does not accommodate all of the values that we need to store in these documents.

The `string` field type has been superseded by the `text` and `keyword` field types [4][5][6].

Since different document types are now stored in different indices, there are no parent/children relationships anymore. As such, I have performed a global rename of `parent` to `testset` and `children` to `results`. Now, to load related content, a `testset_id` field has been added to the `File`, `Settings`, and `Result` models.

Since settings are now in a separate index instead of a nested object on the `TestSet` model, a `load_settings` method has been added to the `TestSet` model.

Other changes:
- Rename `create-index` and `populate-index` to `create-indices` and `populate-indices`
 - Rename `ResultClient` to `Importer`
 - Fix  `DeprecationWarning: invalid escape sequence \m` by prefixing the strings with `r` to convert them to raw strings
 - Fix modal download links on the instance view
 - Add some helpful logging
 - Update README

[1] https://www.elastic.co/guide/en/elasticsearch/reference/7.17/removal-of-types.html#_index_per_document_type
[2] https://www.elastic.co/docs/reference/elasticsearch/mapping-reference/ignore-malformed
[3] https://www.elastic.co/docs/reference/elasticsearch/index-settings/mapping-limit
[4] https://www.elastic.co/blog/strings-are-dead-long-live-strings
[5] https://www.elastic.co/docs/reference/elasticsearch/mapping-reference/text
[6] https://www.elastic.co/docs/reference/elasticsearch/mapping-reference/keyword